Suggest change in wording
The word "value" seems to work better here, and you already used that term a page earlier in the phrase "create a type Visibility that has three possible values". This change improves consistency.  Later in the document you also refer to these things as tags, but that would not be as easy to understand at this point in the document.