deemphasize notes about configuring editor
A friend of mine was trying to learn Elm, but she ended up spending a
bunch of time struggling to set up her editor! Way better to just get
into the book and use /try

Eventually hope to make a very simple editor plugin that is safe to
recommend in this context. No fancy features. Point is just to get
syntax highlighting for people without them needing to spend much time
on it.