move installation way later in the guide
Folks can get to this point with just the interactive examples and the
online editor.

The theory is that you actually lose a decent number of people by
making them start with installation. If they run into trouble, they may
just quit. If they run into trouble but get through, they are feeling
annoyed before they even get started learning!

Doing it this way also feels more "right" in the sense that you are
only asking people to install things once they have a pretty good idea
of what they are getting themselves into.

Finally, the motivation to work through any installation difficulties
will be much higher than if it was the very first page. You have worked
through a bunch of the book, and if you are liking it, this is
something you are actively interested in doing by now.