introduce blue "Edit" buttons to get folks into the online editor
I am hoping that by integrating the button into the content better,
more folks will end up playing with programs in the online editor.

The suggestions of things to try is a way to get folks to read through
the program before reading the description. I hope this will create
more of a mental map for when they start reading the prose.

This change also introduces a simplified diagram to help explain The
Elm Architecture.