just point to the examples
Maybe it is best to have people look at the examples and not point them
towards people who have all decided for themselves to be on one side of
the trade. People who will prefer the other trade may not always find
that they can get their questions answered satisfactorily in that
context.